backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3

# Document parsing
pdfplumber==0.11.4
python-docx==1.1.2

# Embeddings
sentence-transformers==3.2.1
faiss-cpu==1.9.0

# HTTP client
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
